% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig2target.R
\name{sig2target}
\alias{sig2target}
\title{Uses gene signatures, expression data, and dependency data to identify target genes}
\usage{
sig2target(gene.sigs, exp.dat, dep.dat)
}
\arguments{
\item{gene.sigs}{Data frame of gene signatures (each column a unique signature with signature/column name). All gene names must match those in expression data frame.}

\item{exp.dat}{Data frame of gene expression data organized with genes in columns, samples (e.g. cell lines) in rows. Sample name formatting must match that of dependency data (although not all samples must match)}

\item{dep.dat}{Data frame of gene dependency data organized with genes in columns, samples (e.g. cell lines) in rows. Sample name formatting must match that of expression data (although not all samples must match)}
}
\value{
A list 3 objects: 1) Matrix of GSVA scores with each row corresponding to an input signature, 2) data frame of GSVA-vs.-expression Pearson correlations limited to signature genes, 3) data frame of GSVA-vs.-dependency Pearson correlations limited to significantly correlated genes
}
\description{
Runs GSVA using gene signatures and gene expression data, then uses GSVA and dependency data to identify candidate targets.
}
\details{
User inputs data frames of gene signatures, gene expression, and gene dependency data (e.g. from `sigImport`) and function both returns a list of key results and exports those results to .csv files starting with "sig2target".
}
