% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigAccuracy.R
\name{sigAccuracy}
\alias{sigAccuracy}
\title{Calculates how well sample scores predict target gene dependency}
\usage{
sigAccuracy(scores.in, dep.dat, gene, sig.cutoff = -0.5, window = 100)
}
\arguments{
\item{scores.in}{Vector of scores (e.g. GSVA, expression data. etc.) with corresponding sample names (must be formatted like sample names in dependency data set).}

\item{dep.dat}{Data frame of gene dependency data organized with genes in columns, samples (e.g. cell lines) in rows.}

\item{gene}{Name of target gene. Must exactly match that in dependency data set.}

\item{sig.cutoff}{Numerical value that sets the cutoff between "independent" and "dependent" values. Default is -0.5 (DepMap standard).}

\item{window}{Sets the window/chunk size used to scan sample scores and accuracy values that are used in final graph. Larger numbers return "smoother" graph, but less information.}
}
\value{
A list 4 objects: 1) Data frame detailing scores and dependency outcomes for each sample, 2) summary data frame of prediction accuracy (True/false positive/negative), 3) GGplot scatter plot of scores vs dependency data, and 4) GGplot scatter plot of scores (by window) vs predictive accuracy.
}
\description{
Runs linear model between sample scores and dependency values for target gene and uses slope + intercept data to predict dependency values for each sample. Compares predicted and actual dependency values to determine accuracy.
}
\details{
User provides sample scores, target gene, and gene dependency data. Function compares predicted and actual dependencies for each sample and outputs both summary data and graphs. Data frame outputs are also written to files starting with "sigAcc".
}
