% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigImport.R
\name{sigImport}
\alias{sigImport}
\title{Import gene signature(s), gene expression data, and gene dependency data for downstream use}
\usage{
sigImport()
}
\arguments{
\item{NULL}{No parameters passed to function; function prompts user for file names and checks that they are suitable}
}
\value{
A list of 3 objects: 1) a data frame of gene signatures (`gene.sigs`), 2) a data frame of gene expression data (`exp.dat`), and 3) a data frame of gene dependency data (`dep.dat`)
}
\description{
Guides user in importing key data for use in the `sig2target`, `target2sig` and `sigAccuracy` functions from this package. For data formatting, please pay close attention to each prompt!
}
\details{
Function is only compatible with .csv files stored in the current working directory. When prompted, user provides full file names and `sigImport` checks that they are compatible with the requirements of functions in this package.
}
