% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target2sig.R
\name{target2sig}
\alias{target2sig}
\title{Uses target gene dependency data to create/test new gene signature(s)}
\usage{
target2sig(dep.dat, gene, exp.dat, cutoff = 100, gene.sig = NA)
}
\arguments{
\item{dep.dat}{Data frame of gene dependency data organized with genes in columns, samples (e.g. cell lines) in rows. Sample name formatting must match that of expression data (although not all samples must match)}

\item{gene}{Name of target gene. Must exactly match that in dependency data set.}

\item{exp.dat}{Data frame of gene expression data organized with genes in columns, samples (e.g. cell lines) in rows. Sample name formatting must match that of dependency data (although not all samples must match)}

\item{cutoff}{Positive integer value that sets the maximum size of the output gene signature. Default is 100.}

\item{gene.sig}{OPTIONAL. A single gene signature vector. Names must match those in dependency/expression data sets.}
}
\value{
A list 3 objects: 1) List of new gene signature(s), 2) matrix of GSVA scores with each row corresponding to new signature(s), 3) data frame of GSVA-vs.-dependency Pearson correlations limited to significantly correlated genes
}
\description{
Creates new gene signature(s) by finding genes whose expression correlates to target gene dependency. New signature(s) used to run GSVA and then compare back to gene dependency data.
}
\details{
User inputs a target gene name and data frames of gene dependency and gene expression data (e.g. from `sigImport`) and function both returns a list of key results and exports those results to .csv files starting with "target2sig".
}
